+++ src/main/java/com/acon/server/member/application/mapper/RecentGuidedSpotMapper.java
@@ -1,7 +1,7 @@
 package com.acon.server.member.application.mapper;
 
 import com.acon.server.member.domain.entity.RecentGuidedSpot;
-import com.acon.server.member.infra.entity.RecentGuidedSpotEntity;
+import com.acon.server.member.infra.entity.member.RecentGuidedSpotEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
