+++ src/main/java/com/acon/server/member/infra/entity/review/ReviewEntity.java
@@ -1,4 +1,4 @@
-package com.acon.server.member.infra.entity;
+package com.acon.server.member.infra.entity.review;
 
 import com.acon.server.global.entity.BaseTimeEntity;
 import jakarta.persistence.Column;
@@ -7,7 +7,6 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import jakarta.persistence.UniqueConstraint;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -16,32 +15,31 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(
-        name = "recent_viewed_spot", uniqueConstraints = @UniqueConstraint(
-                name = "unique_recent_viewed_spot_member_id_spot_id", columnNames = {"member_id", "spot_id"}
-        )
-
-)
-public class RecentViewedSpotEntity extends BaseTimeEntity {
+@Table(name = "review")
+public class ReviewEntity extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(name = "member_id", nullable = false)
-    private Long memberId;
-
     @Column(name = "spot_id", nullable = false)
     private Long spotId;
 
+    @Column(name = "member_id", nullable = false)
+    private Long memberId;
+
+    @Column(name = "acorn_count", nullable = false, columnDefinition = "integer default 0")
+    private int acornCount;
+
+    @Column(name = "local_acorn", nullable = false, columnDefinition = "boolean default false")
+    private boolean localAcorn;
+
     @Builder
-    public RecentViewedSpotEntity(
-            Long id,
-            Long memberId,
-            Long spotId
-    ) {
+    public ReviewEntity(Long id, Long spotId, Long memberId, int acornCount, boolean localAcorn) {
         this.id = id;
-        this.memberId = memberId;
         this.spotId = spotId;
+        this.memberId = memberId;
+        this.acornCount = acornCount;
+        this.localAcorn = localAcorn;
     }
 }
